pyproject.toml
[tool.poetry]
name = "kebede-bot"
version = "0.1.0"
description = ""
authors = ["jabez <jabezkassa2022@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
